Fixed test suite failures by refactoring
This commit contains refactoring done to the test suite to better accommodate the refactoring done to the api_client.

- Addressed Static Type Errors in the Request Class
On api_client_test, the type checker complained Attribute "filters" is unknown because attributes were only being assigned in __init__ without being formally declared. I redefined the Request class to be a dataclass to resolve the errors related to Request

- Corrected the Test Mocking Strategy (test_do & test_get_entity)
Refactoring led to creating the Client instance in the setup method, and to manually patch the attribute on that instance: self.client.http_client = MagicMock(...). This ensures every test gets a fresh mock on a real client object.

This change introduced a few static type errors, for this I used typing.cast inside each test method.

- Refined the Test Assertions for Logical Correctness
Due to refactoring, the _do method doesn't just send the request it receives. It first calls prepare_request() and sends the new, prepared request object, and it also adds a timeout keyword argument. Because of this I updated mock_http_client.send.assert_called_once_with(...) to check for the correct prepared_request object and the timeout value.