Continuation of SEH Implementation
This commit contains SEH implementation into the batch.py file.

Further, it contains a new addition of exceptions.py, where exceptions are declared and handled, so they can be called from a single file as best practice.

This is only one, of a series of commits I'll be doing to implement SEH on the SDK. My reason for breaking them up into batches is to keep each commit relatively small and easy to review.