SEH Implementation Part 2
This commit is batch 2 of the SEH I am implementing to the SDK. In this commit, the following files were worked on:

1. editory.py
2. event.py
3. headers.py

Following, I'll be breaking down the areas of improvement that were identified on the files:

editor.py

The from_json method used direct key access (data['key']) for every single field. If the API response were missing any of those fields for any reason, the entire method will crash with a KeyError.

Previously, datetime.fromisoformat(data['dateStarted']) would crash if the date string was not in a valid ISO 8601 format.

The to_json method used to call editor.date_started.isoformat(). Since the __init__ method allowed date_started to be None, this used to raise an AttributeError if we tried to serialize an Editor object where that date isn't set.

event.py

As seen before, the from_json method used direct data['key'] access, which would crash if any key is missing.

Both dateCreated and datePublished were parsed with fromisoformat without any error handling.

The to_json method would crash if date_created or date_published are None, as it tried to call .isoformat() on them directly.

The event_type field was accepted as any string. If the API sent an unexpected value, the error would be silently accepted here, which could cause a crash down the line.

headers.py

Finally, again, the from_json method used direct data['key'] access, which would crash if any header were not present in the input data.

The to_json method would crash by calling .isoformat() on last_modified if that attribute is None.

According to universal web standards, the Last-Modified header in HTTP does not use the ISO 8601 format but instead uses the RFC 2822 format. Calling datetime.fromisoformat() on this string would fail with a ValueError.