Implemented SDK logging capability and commit feedback
This commit mainly focuses on adding logging support into the SDK, so it can handle the following events, providing more clarity to the user should something go wrong:

- HTTP status errors

- Request errors

- JSON decoding errors

- Chunk download failures

This commit also includes a new example, under example > logging, which showcases the new logging capabilities. Usage and documentation for the script can be found within it.

Finally, I also included the feedback proposed in the PR.